requirements.txt
django
djangorestframework
rest_framework_simplejwt
python-dotenv